% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02C_answer_filter.R
\name{answer_filter_is}
\alias{answer_filter_is}
\title{获取答案级黑名单里的精确匹配}
\usage{
answer_filter_is(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
获取答案级黑名单里的精确匹配
}
\examples{
answer_filter_is();
}
