% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02C_answer_filter.R
\name{answer_filter_like}
\alias{answer_filter_like}
\title{答案级黑名单模糊匹配}
\usage{
answer_filter_like(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
答案级黑名单模糊匹配
}
\examples{
answer_filter_like();
}
