% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02D_regexp_rule_filter.R
\name{del_car_VIN_ques}
\alias{del_car_VIN_ques}
\title{车架在问题在的应用}
\usage{
del_car_VIN_ques(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
车架在问题在的应用
}
\examples{
del_car_VIN_ques();
}
