% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02D_regexp_rule_filter.R
\name{del_mobile_answ}
\alias{del_mobile_answ}
\title{删除答案中的电话号码}
\usage{
del_mobile_answ(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
删除答案中的电话号码
}
\examples{
del_mobile_answ();
}
