% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02D_regexp_rule_filter.R
\name{del_orderNumber_answ}
\alias{del_orderNumber_answ}
\title{删除答案中的订单豪华版}
\usage{
del_orderNumber_answ(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
删除答案中的订单豪华版
}
\examples{
del_orderNumber_answ()
}
