% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02D_regexp_rule_filter.R
\name{del_orderNumber_ques}
\alias{del_orderNumber_ques}
\title{删除问题中的订单信息}
\usage{
del_orderNumber_ques(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
删除问题中的订单信息
}
\examples{
del_orderNumber_ques();
}
