% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02A_file_filter.R
\name{file_filter_is}
\alias{file_filter_is}
\title{获取文件黑名单里的精确匹配}
\usage{
file_filter_is(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
获取文件黑名单里的精确匹配
}
\examples{
file_filter_is();
}
