% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02A_file_filter.R
\name{file_filter_like}
\alias{file_filter_like}
\title{文件级黑名单模糊匹配}
\usage{
file_filter_like(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
文件级黑名单模糊匹配
}
\examples{
file_filter_like();
}
