% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_filter_apply.R
\name{item_answ_version_filtered}
\alias{item_answ_version_filtered}
\title{删除版本化管理}
\usage{
item_answ_version_filtered(brand = "JBLH", version = "V1")
}
\arguments{
\item{brand}{品牌}

\item{version}{版本}
}
\value{
返回值
}
\description{
删除版本化管理
}
\examples{
item_answ_version_filtered();
}
