% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_filter_apply.R
\name{item_ques_version_filtered}
\alias{item_ques_version_filtered}
\title{将版本进行过滤操作}
\usage{
item_ques_version_filtered(brand = "JBLH", version = "V1")
}
\arguments{
\item{brand}{品牌}

\item{version}{版本}
}
\value{
返回值
}
\description{
将版本进行过滤操作
}
\examples{
item_ques_version_filtered();
}
