% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_keyWord_filter.R
\name{keyword_filter_is}
\alias{keyword_filter_is}
\title{针对需要删除的关键词进行定义}
\usage{
keyword_filter_is(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
针对需要删除的关键词进行定义
}
\examples{
keyword_filter_is();
}
