% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_keyWord_filter.R
\name{nsim_item_answ_version}
\alias{nsim_item_answ_version}
\title{处理答案版本化问题}
\usage{
nsim_item_answ_version(brand = "JBLH", version = "V1")
}
\arguments{
\item{brand}{品牌}

\item{version}{版本}
}
\value{
返回值
}
\description{
处理答案版本化问题
}
\examples{
nsim_item_answ_version();
}
