% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_keyWord_filter.R
\name{nsim_item_ques_version}
\alias{nsim_item_ques_version}
\title{将问题进行版本化处理}
\usage{
nsim_item_ques_version(brand = "JBLH", version = "V1")
}
\arguments{
\item{brand}{品牌}

\item{version}{版本}
}
\value{
返回值
}
\description{
将问题进行版本化处理
}
\examples{
nsim_item_ques_version();
}
