% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_filter_apply.R
\name{nsim_nsbl_version}
\alias{nsim_nsbl_version}
\title{提供主料的品牌处理}
\usage{
nsim_nsbl_version(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
提供主料的品牌处理
}
\examples{
nsim_nsbl_version();
}
