% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_keyWord_filter.R
\name{nsim_nscl_version}
\alias{nsim_nscl_version}
\title{通过对语料进行删词操作，产生一个新的版本}
\usage{
nsim_nscl_version(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
通过对语料进行删词操作，产生一个新的版本
}
\examples{
nsim_nscl_version();
}
