% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02B_question_filter.R
\name{question_filter_is}
\alias{question_filter_is}
\title{获取问题级黑名单里的精确匹配}
\usage{
question_filter_is(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
获取问题级黑名单里的精确匹配
}
\examples{
question_filter_is();
}
