% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02B_question_filter.R
\name{question_filter_like}
\alias{question_filter_like}
\title{问题级黑名单模糊匹配}
\usage{
question_filter_like(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
问题级黑名单模糊匹配
}
\examples{
question_filter_like();
}
